Revamp early colonist logic (especially for Rome & Carthage)
Give AI Carthage free harbor on founding on that city.

Revamped the early colonists logic:
a) give a message when it triggers, using an immersive description
b) combined the foundation of Carthage event into the early colonists logic
c) give Carthage an extra colonist, directly in southern Iberia
d) give Rome a generic version of the event, as it doesn't need to have the galley placed near Iberia for it to target a sensible location for a new city
e) added a bit of randomness to the exact dates, similarly to how it was done for the Carthage foundation event
f) moved the wine in southern Iberia 1 tile east of Gades/Gadir, as it is the mostly likely city to be founded. This shouldn't affect Cordoba for later years, or the 600/1700AD spawn scenarios.
g) add massive value to 2 city locations in Iberia for Carthage, otherwise it will refuse to settle there.
h) Reduce the settler value to 1 in the westernmost North African plots (Mauritania/Morocco) for Carthage