Add two indy cities with Norse names if Rus doesn't spawn
Rus may not spawn because of its "limited" impact rating, but we still want some indy cities to fill that void.
Norse names chosen as it is perhaps the least controversial to Western audiences. I understand that some Slav people consider the "Norman Hypothesis" to undermine their national history, but the names will disappear when Russia or Poland conquers the cities anyhow.