Persia balance changes
Re-introduced Greek colonization waves; this is mostly so that the first wave can settle Byzantium, and subsequent waves can target places in the Black Sea. Previously, since they all spawned at once, none of them targeted the Black Sea coast.
Persia gets Contract on start, so it immediately has the optional to vassalize other civs
Persia no longer gets the "you are too powerful for us" refusal for vassalization
The whole point of their UHV is to give them a big perceived power boost for getting vassals, so this defeats the point!
Removed condition that would prevent a civ capitulating to another civ that was birth protected
Just because a civ is birth protected doesn't mean another civ shouldn't become a vassal!
If there are any edge cases to this for specific civs that are meant to only conquer and not subjugate, I will add civ-specific exemptions
Civs have less strict denial of capitulation conditions in the ancient and classical eras
"average power" used to calculate if a civ should capitulate are less reliable in the ancient and classical eras, when there are a lot less civs around
Greek city-states name takes preference over Empire name if they have the Republic civic.
A couple tiles in Thrace between Pella and Byzantium are no longer on the Greek settler map
Persian maintenance modifiers buffed (frequently running into negative treasury at 0% due to the number of low population cities)
Better condition for switch from Cyrus to Darius leader for Persia (used to be "on empire", which is triggered quite quickly for Persia)
If Assyria and Babylonia are dead, switch to Darius
Persia UHV tweaks (less total population percent, include vassals in population percent count, less wonders, include vassals in wonder count)