Give 'inferior' heavy swordsman UU to Romans & Byzantines
It is available earlier in the tech tree but has 1 less strength than a default heavy swordsman. This is to reflect the relative lowering of quality of late roman / byzantine heavy infantry.
Buff to Roman city maintenance modifiers.
Automatically upgrade all legions to 0-experience Comitatenses if Rome is AI.
Modify Roman and Byzantine tech preferences.